requirements.txt
torch>=1.9.0
transformers>=4.10.0
datasets>=1.11.0
requests>=2.25.1
tqdm>=4.62.0
numpy>=1.21.0
scikit-learn>=0.24.2
google-cloud-aiplatform>=1.24.0
pandas>=1.3.0
rouge_score>=0.1.2
python-dotenv>=0.19.0
plotly>=5.3.1
flask
flask-bcrypt
firebase-admin
diffusers>=0.11.1
torchvision>=0.10.0
Pillow>=8.3.1
accelerate>=0.12.0